Whoops. :)

+++ include/SDL_stdinc.h
@@ -134,7 +134,7 @@ typedef uint64_t Uint64;
 #else
 /* This is really just a hack to prevent the compiler from complaining */
 typdef Sint32 Sint64;
-typdef Uint32 Uint32;
+typdef Uint32 Uint64;
 #endif
 
 /* Make sure the types really have the right sizes */
